+++ examples/multiFileProject/multiFileProject.cpp
@@ -1,0 +1,13 @@
+/****************************************************************************************************************************
+  multiFileProject.cpp
+  
+  For ESP8266 boards
+  Written by Khoi Hoang
+
+  Built by Khoi Hoang https://github.com/khoih-prog/ESP8266_PWM
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#include "multiFileProject.h"