+++ src/ESP8266_PWM_ISR.hpp
@@ -29,329 +29,220 @@
 #ifndef PWM_ISR_GENERIC_HPP
 #define PWM_ISR_GENERIC_HPP
 
-#include <string.h>
+#if !defined(ESP8266)
+  #error This code is designed to run on ESP8266 and ESP8266-based boards! Please check your Tools->Board setting.
+#endif
 
-/////////////////////////////////////////////////// 
+#ifndef ESP8266_PWM_VERSION
+  #define ESP8266_PWM_VERSION           "ESP8266_PWM v1.2.0"
+  
+  #define ESP8266_PWM_VERSION_MAJOR     1
+  #define ESP8266_PWM_VERSION_MINOR     2
+  #define ESP8266_PWM_VERSION_PATCH     0
 
+  #define ESP8266_PWM_VERSION_INT       1002000
+#endif
 
-uint64_t timeNow()
-{
-#if USING_MICROS_RESOLUTION  
-  return ( (uint64_t) micros() );
+#ifndef TIMER_INTERRUPT_DEBUG
+  #define TIMER_INTERRUPT_DEBUG      0
+#endif
+
+#ifndef _PWM_LOGLEVEL_
+  #define _PWM_LOGLEVEL_       1
+#endif
+
+#include "PWM_Generic_Debug.h"
+
+#define CONFIG_ESP32_APPTRACE_ENABLE
+
+#include <stddef.h>
+
+#ifdef ESP8266
+  extern "C"
+  {
+    #include "ets_sys.h"
+    #include "os_type.h"
+    #include "mem.h"
+  }
 #else
-  return ( (uint64_t) millis() );
-#endif    
-}
-  
-/////////////////////////////////////////////////// 
+  #include <inttypes.h>
+#endif
 
-ESP8266_PWM_ISR::ESP8266_PWM_ISR()
-  : numChannels (-1)
+#if defined(ARDUINO)
+  #if ARDUINO >= 100
+    #include <Arduino.h>
+  #else
+    #include <WProgram.h>
+  #endif
+#endif
+
+#define ESP8266_PWM_ISR ESP8266_PWM
+
+typedef void (*esp8266_timer_callback)();
+typedef void (*esp8266_timer_callback_p)(void *);
+
+#if !defined(USING_MICROS_RESOLUTION)
+  #warning Not USING_MICROS_RESOLUTION, using millis resolution
+  #define USING_MICROS_RESOLUTION       false
+#endif
+
+class ESP8266_PWM_ISR 
 {
-}
 
-///////////////////////////////////////////////////
+  public:
+    // maximum number of PWM channels
+#define MAX_NUMBER_CHANNELS        16
 
-void ESP8266_PWM_ISR::init() 
-{
-  uint64_t currentTime = timeNow();
-   
-  for (uint8_t channelNum = 0; channelNum < MAX_NUMBER_CHANNELS; channelNum++) 
-  {
-    memset((void*) &PWM[channelNum], 0, sizeof (PWM_t));
-    PWM[channelNum].prevTime = currentTime;    
-    PWM[channelNum].pin      = INVALID_ESP8266_PIN;
-  }
-  
-  numChannels = 0;
-}
+    // constructor
+    ESP8266_PWM_ISR();
 
-///////////////////////////////////////////////////
+    void init();
 
-void IRAM_ATTR ESP8266_PWM_ISR::run() 
-{    
-  uint64_t currentTime = timeNow();
-
-  for (uint8_t channelNum = 0; channelNum < MAX_NUMBER_CHANNELS; channelNum++) 
-  {
-    // If enabled => check
-    // start period / dutyCycle => digitalWrite HIGH
-    // end dutyCycle =>  digitalWrite LOW
-    if (PWM[channelNum].enabled) 
-    {
-      if ( (uint32_t) (currentTime - PWM[channelNum].prevTime) <= PWM[channelNum].onTime ) 
-      {              
-        if (!PWM[channelNum].pinHigh)
-        {
-          digitalWrite(PWM[channelNum].pin, HIGH);
-          PWM[channelNum].pinHigh = true;
-          
-          // callback when PWM pulse starts (HIGH)
-          if (PWM[channelNum].callbackStart != nullptr)
-          {
-            (*(esp8266_timer_callback) PWM[channelNum].callbackStart)();
-          }
-        }
-      }
-      else if ( (uint32_t) (currentTime - PWM[channelNum].prevTime) < PWM[channelNum].period ) 
-      {
-        if (PWM[channelNum].pinHigh)
-        {
-          digitalWrite(PWM[channelNum].pin, LOW);
-          PWM[channelNum].pinHigh = false;
-          
-          // callback when PWM pulse stops (LOW)
-          if (PWM[channelNum].callbackStop != nullptr)
-          {
-            (*(esp8266_timer_callback) PWM[channelNum].callbackStop)();
-          }
-        }
-      }
-      //else 
-      else if ( (uint32_t) (currentTime - PWM[channelNum].prevTime) >= PWM[channelNum].period )   
-      {
-        PWM[channelNum].prevTime = currentTime;
-      }      
-    }
-  }
-}
-
-
-///////////////////////////////////////////////////
-
-// find the first available slot
-// return -1 if none found
-int ESP8266_PWM_ISR::findFirstFreeSlot() 
-{
-  // all slots are used
-  if (numChannels >= MAX_NUMBER_CHANNELS) 
-  {
-    return -1;
-  }
-
-  // return the first slot with zero period and not enabled (i.e. free)
-  for (uint8_t channelNum = 0; channelNum < MAX_NUMBER_CHANNELS; channelNum++) 
-  {
-    if ( (PWM[channelNum].period == 0) && !PWM[channelNum].enabled )
-    {
-      return channelNum;
-    }
-  }
-
-  // no free slots found
-  return -1;
-}
-
-///////////////////////////////////////////////////
-
-// Return the channelNum if OK, -1 if error
-int ESP8266_PWM_ISR::setupPWMChannel(uint32_t pin, uint32_t period, uint32_t dutycycle, void* cbStartFunc, void* cbStopFunc)
-{
-  int channelNum;
-  // Invalid input, such as period = 0, etc
-  if ( (period == 0) || (dutycycle > 100) )
-  {
-    PWM_LOGERROR("Error: Invalid period or dutycycle");
-    return -1;
-  }
-
-  if (numChannels < 0) 
-  {
-    init();
-  }
- 
-  channelNum = findFirstFreeSlot();
-  
-  if (channelNum < 0) 
-  {
-    return -1;
-  }
-
-  PWM[channelNum].pin           = pin;
-  PWM[channelNum].period        = period;
-  PWM[channelNum].onTime        = ( period * dutycycle ) / 100;
-  
-  pinMode(pin, OUTPUT);
-  digitalWrite(pin, HIGH);
-  PWM[channelNum].pinHigh       = true;
-  
-  PWM[channelNum].prevTime      = timeNow();
-  
-  PWM[channelNum].callbackStart = cbStartFunc;
-  PWM[channelNum].callbackStop  = cbStopFunc;
-  
-  PWM_LOGDEBUG0("Channel : "); PWM_LOGDEBUG0(channelNum); PWM_LOGDEBUG0("\tPeriod : "); PWM_LOGDEBUG0(PWM[channelNum].period);
-  PWM_LOGDEBUG0("\t\tOnTime : "); PWM_LOGDEBUG0(PWM[channelNum].onTime); PWM_LOGDEBUG0("\tStart_Time : "); PWM_LOGDEBUGLN0(PWM[channelNum].prevTime);
- 
-  numChannels++;
-  
-  PWM[channelNum].enabled      = true;
-  
-  return channelNum;
-}
-
-///////////////////////////////////////////////////
-
-bool ESP8266_PWM_ISR::modifyPWMChannel_Period(unsigned channelNum, uint32_t pin, uint32_t period, uint32_t dutycycle)
-{
-  // Invalid input, such as period = 0, etc
-  if ( (period == 0) || (dutycycle > 100) )
-  {
-    PWM_LOGERROR("Error: Invalid period or dutycycle");
-    return false;
-  }
-
-  if (channelNum > MAX_NUMBER_CHANNELS) 
-  {
-    PWM_LOGERROR("Error: channelNum > MAX_NUMBER_CHANNELS");
-    return false;
-  }
-  
-  if (PWM[channelNum].pin != pin) 
-  {
-    PWM_LOGERROR("Error: channelNum and pin mismatched");
-    return false;
-  }
-
-  PWM[channelNum].period        = period;
-  PWM[channelNum].onTime        = ( period * dutycycle ) / 100;
-  
-  digitalWrite(pin, HIGH);
-  PWM[channelNum].pinHigh       = true;
-  
-  PWM[channelNum].prevTime      = timeNow();
-   
-  PWM_LOGDEBUG0("Channel : "); PWM_LOGDEBUG0(channelNum); PWM_LOGDEBUG0("\tPeriod : "); PWM_LOGDEBUG0(PWM[channelNum].period);
-  PWM_LOGDEBUG0("\t\tOnTime : "); PWM_LOGDEBUG0(PWM[channelNum].onTime); PWM_LOGDEBUG0("\tStart_Time : "); PWM_LOGDEBUGLN0(PWM[channelNum].prevTime);
-  
-  return true;
-}
-
-
-///////////////////////////////////////////////////
-
-void ESP8266_PWM_ISR::deleteChannel(unsigned channelNum) 
-{
-  if (channelNum >= MAX_NUMBER_CHANNELS) 
-  {
-    return;
-  }
-
-  // nothing to delete if no timers are in use
-  if (numChannels == 0) 
-  {
-    return;
-  }
-
-  // don't decrease the number of timers if the specified slot is already empty (zero period, invalid)
-  if ( (PWM[channelNum].pin != INVALID_ESP8266_PIN) && (PWM[channelNum].period != 0) )
-  {
-    memset((void*) &PWM[channelNum], 0, sizeof (PWM_t));
+    // this function must be called inside loop()
+    void IRAM_ATTR run();
     
-    PWM[channelNum].pin = INVALID_ESP8266_PIN;
+    //////////////////////////////////////////////////////////////////
+    // PWM
+    // Return the channelNum if OK, -1 if error
+    int setPWM(const uint32_t& pin, const double& frequency, const double& dutycycle, esp8266_timer_callback StartCallback = nullptr, 
+                esp8266_timer_callback StopCallback = nullptr)
+    {
+      double period = 0.0;
+      
+      if ( ( frequency > 0 ) && ( frequency <= 500 ) )
+      {
+#if USING_MICROS_RESOLUTION
+      // period in us
+      period = 1000000.0f / frequency;
+#else    
+      // period in ms
+      period = 1000.0f / frequency;
+#endif
+      }
+      else
+      {       
+        PWM_LOGERROR("Error: Invalid frequency, max is 500Hz");
+        
+        return -1;
+      }
+      
+      return setupPWMChannel(pin, period, dutycycle, (void *) StartCallback, (void *) StopCallback);      
+    }
+
+    // period in us
+    // Return the channelNum if OK, -1 if error
+    int setPWM_Period(const uint32_t& pin, const double& period, const double& dutycycle, esp8266_timer_callback StartCallback = nullptr,
+                       esp8266_timer_callback StopCallback = nullptr)  
+    {     
+      return setupPWMChannel(pin, period, dutycycle, (void *) StartCallback, (void *) StopCallback);      
+    }    
     
-    // update number of timers
-    numChannels--;
-  }
-}
-
-///////////////////////////////////////////////////
-
-void ESP8266_PWM_ISR::restartChannel(unsigned channelNum) 
-{
-  if (channelNum >= MAX_NUMBER_CHANNELS) 
-  {
-    return;
-  }
-
-  PWM[channelNum].prevTime = timeNow();
-}
-
-///////////////////////////////////////////////////
-
-bool ESP8266_PWM_ISR::isEnabled(unsigned channelNum) 
-{
-  if (channelNum >= MAX_NUMBER_CHANNELS) 
-  {
-    return false;
-  }
-
-  return PWM[channelNum].enabled;
-}
-
-///////////////////////////////////////////////////
-
-void ESP8266_PWM_ISR::enable(unsigned channelNum) 
-{
-  if (channelNum >= MAX_NUMBER_CHANNELS) 
-  {
-    return;
-  }
-
-  PWM[channelNum].enabled = true;
-}
-
-///////////////////////////////////////////////////
-
-void ESP8266_PWM_ISR::disable(unsigned channelNum) 
-{
-  if (channelNum >= MAX_NUMBER_CHANNELS) 
-  {
-    return;
-  }
-
-  PWM[channelNum].enabled = false;
-}
-
-///////////////////////////////////////////////////
-
-void ESP8266_PWM_ISR::enableAll() 
-{
-  // Enable all timers with a callback assigned (used)
-
-  for (uint8_t channelNum = 0; channelNum < MAX_NUMBER_CHANNELS; channelNum++) 
-  {
-    if (PWM[channelNum].period != 0)
+    //////////////////////////////////////////////////////////////////
+    
+    // low level function to modify a PWM channel
+    // returns the true on success or false on failure
+    bool modifyPWMChannel(const unsigned& channelNum, const uint32_t& pin, const double& frequency, const double& dutycycle)
     {
-      PWM[channelNum].enabled = true;
+      uint32_t period = 0;
+      
+      if ( ( frequency > 0 ) && ( frequency <= 500 ) )
+      {
+#if USING_MICROS_RESOLUTION
+      // period in us
+      period = 1000000.0f / frequency;
+#else    
+      // period in ms
+      period = 1000.0f / frequency;
+#endif
+      }
+      else
+      {       
+        PWM_LOGERROR("Error: Invalid frequency, max is 500Hz");
+        return false;
+      }
+      
+      return modifyPWMChannel_Period(channelNum, pin, period, dutycycle);
     }
-  }
-}
+    
+    //period in us
+    bool modifyPWMChannel_Period(const unsigned& channelNum, const uint32_t& pin, const double& period, const double& dutycycle);
 
-///////////////////////////////////////////////////
+    // destroy the specified PWM channel
+    void deleteChannel(const unsigned& channelNum);
 
-void ESP8266_PWM_ISR::disableAll() 
-{
-  // Disable all timers with a callback assigned (used)
+    // restart the specified PWM channel
+    void restartChannel(const unsigned& channelNum);
 
-  for (uint8_t channelNum = 0; channelNum < MAX_NUMBER_CHANNELS; channelNum++) 
-  {
-    if (PWM[channelNum].period != 0)
+    // returns true if the specified PWM channel is enabled
+    bool isEnabled(const unsigned& channelNum);
+
+    // enables the specified PWM channel
+    void enable(const unsigned& channelNum);
+
+    // disables the specified PWM channel
+    void disable(const unsigned& channelNum);
+
+    // enables all PWM channels
+    void enableAll();
+
+    // disables all PWM channels
+    void disableAll();
+
+    // enables the specified PWM channel if it's currently disabled, and vice-versa
+    void toggle(const unsigned& channelNum);
+
+    // returns the number of used PWM channels
+    unsigned getnumChannels();
+
+    // returns the number of available PWM channels
+    unsigned getNumAvailablePWMChannels() 
     {
-      PWM[channelNum].enabled = false;
-    }
-  }
-}
+      return MAX_NUMBER_CHANNELS - numChannels;
+    };
 
-///////////////////////////////////////////////////
+  private:
 
-void ESP8266_PWM_ISR::toggle(unsigned channelNum) 
-{
-  if (channelNum >= MAX_NUMBER_CHANNELS) 
-  {
-    return;
-  }
+    // low level function to initialize and enable a new PWM channel
+    // returns the PWM channel number (channelNum) on success or
+    // -1 on failure (f == NULL) or no free PWM channels 
+    int setupPWMChannel(const uint32_t& pin, const double& period, const double& dutycycle, void* cbStartFunc = nullptr, void* cbStopFunc = nullptr);
+    
 
-  PWM[channelNum].enabled = !PWM[channelNum].enabled;
-}
+    // find the first available slot
+    int findFirstFreeSlot();
 
-///////////////////////////////////////////////////
+    typedef struct 
+    {
+      ///////////////////////////////////
+      
+      
+      ///////////////////////////////////
+      
+      uint64_t      prevTime;           // value returned by the micros() or millis() function in the previous run() call
+      double        period;             // period value, in us / ms
+      uint32_t      onTime;             // onTime value, ( period * dutyCycle / 100 ) us  / ms
+      
+      void*         callbackStart;      // pointer to the callback function when PWM pulse starts (HIGH)
+      void*         callbackStop;       // pointer to the callback function when PWM pulse stops (LOW)
+      
+      ////////////////////////////////////////////////////////////
+      
+      uint32_t      pin;                // PWM pin
+      bool          pinHigh;            // true if PWM pin is HIGH
+      ////////////////////////////////////////////////////////////
+      
+      bool          enabled;            // true if enabled
+    } PWM_t;
 
-unsigned ESP8266_PWM_ISR::getnumChannels() 
-{
-  return numChannels;
-}
+    volatile PWM_t PWM[MAX_NUMBER_CHANNELS];
+
+    // actual number of PWM channels in use (-1 means uninitialized)
+    volatile int numChannels;
+};
 
 
-#endif    // PWM_ISR_GENERIC_HPP
+//#include "ESP8266_PWM_ISR.hpp"
+
+#endif    // PWM_ISR_GENERIC_H
+